working version of vf2 mdio

+++ include/memory/MemoryMap.hpp
@@ -7,7 +7,7 @@ struct MemoryMap {
     };
 
     struct App : Entry {
-        uintptr_t main;
+        unsigned long main;
         Entry text;
         Entry data;
         Entry rodata;
